changes balance to float and refactors tests

+++ db/migrate/20210224041142_create_payment_accounts.rb
@@ -2,7 +2,7 @@ class CreatePaymentAccounts < ActiveRecord::Migration
   def change
     create_table :payment_accounts do |t|
       t.references :user, index: true, foreign_key: true, null: false
-      t.integer :balance, null: false, default: 0
+      t.float :balance, null: false, default: 0
       t.timestamps null: false
     end
   end
